requirements.txt
beautifulsoup4
absl-py

black==20.8b1
flake8
pre-commit
pylint